holiday item refactoring
